# spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "rustygene-api",
    "description": "",
    "contact": {
      "name": "rustygiene contributors"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/citations": {
      "get": {
        "tags": [],
        "operationId": "citations_list",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "citations_create",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [],
        "operationId": "events_list",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "events_create",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/v1/families": {
      "get": {
        "tags": [],
        "operationId": "families_list",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "families_create",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/v1/media": {
      "get": {
        "tags": [],
        "operationId": "media_list",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "media_upload",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/v1/notes": {
      "get": {
        "tags": [],
        "operationId": "notes_list",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "notes_create",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/v1/persons": {
      "get": {
        "tags": [],
        "operationId": "persons_list",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "persons_create",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/v1/repositories": {
      "get": {
        "tags": [],
        "operationId": "repositories_list",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "repositories_create",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/v1/research-log": {
      "get": {
        "tags": [],
        "operationId": "research_log_list",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "research_log_create",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/v1/sources": {
      "get": {
        "tags": [],
        "operationId": "sources_list",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "sources_create",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/v1/staging": {
      "get": {
        "tags": [],
        "operationId": "staging_list",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "staging_submit",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
  },
...
}